+++ x/bank/keeper/view.go
@@ -134,12 +134,15 @@ func (k BaseViewKeeper) GetAccountsBalances(ctx context.Context) []types.Balance
 
 // GetBalance returns the balance of a specific denomination for a given account
 // by address.
-func (k BaseViewKeeper) GetBalance(ctx context.Context, addr sdk.AccAddress, denom string) sdk.Coin {
-	amt, err := k.Balances.Get(ctx, collections.Join(addr, denom))
+func (k BaseViewKeeper) GetBalance(ctx sdk.Context, addr sdk.AccAddress, denom string) sdk.Coin {
+	accountStore := k.getAccountStore(ctx, addr)
+	bz := accountStore.Get([]byte(denom))
+	balance, err := UnmarshalBalanceCompat(k.cdc, bz, denom)
 	if err != nil {
-		return sdk.NewCoin(denom, math.ZeroInt())
+		panic(err)
 	}
-	return sdk.NewCoin(denom, amt)
+
+	return balance
 }
 
 // IterateAccountBalances iterates over the balances of a single account and
@@ -152,12 +155,13 @@ func (k BaseViewKeeper) IterateAccountBalances(ctx sdk.Context, addr sdk.AccAddr
 	defer iterator.Close()
 
 	for ; iterator.Valid(); iterator.Next() {
-		var amount math.Int
-		if err := amount.Unmarshal(iterator.Value()); err != nil {
+		denom := string(iterator.Key())
+		balance, err := UnmarshalBalanceCompat(k.cdc, iterator.Value(), denom)
+		if err != nil {
 			panic(err)
 		}
 
-		if cb(sdk.NewCoin(string(iterator.Key()), amount)) {
+		if cb(balance) {
 			break
 		}
 	}
@@ -182,12 +186,12 @@ func (k BaseViewKeeper) IterateAllBalances(ctx sdk.Context, cb func(sdk.AccAddre
 			panic(err)
 		}
 
-		var amount math.Int
-		if err := amount.Unmarshal(iterator.Value()); err != nil {
+		balance, err := UnmarshalBalanceCompat(k.cdc, iterator.Value(), denom)
+		if err != nil {
 			panic(err)
 		}
 
-		if cb(address, sdk.NewCoin(denom, amount)) {
+		if cb(address, balance) {
 			break
 		}
 	}
@@ -275,4 +279,36 @@ func (k BaseViewKeeper) ValidateBalance(ctx context.Context, addr sdk.AccAddress
 	}
 
 	return nil
+}
+
+// getAccountStore gets the account store of the given address.
+func (k BaseViewKeeper) getAccountStore(ctx sdk.Context, addr sdk.AccAddress) prefix.Store {
+	store := ctx.KVStore(k.storeKey)
+	return prefix.NewStore(store, types.CreateAccountBalancesPrefix(addr))
+}
+
+// getDenomAddressPrefixStore returns a prefix store that acts as a reverse index
+// between a denomination and account balance for that denomination.
+func (k BaseViewKeeper) getDenomAddressPrefixStore(ctx sdk.Context, denom string) prefix.Store {
+	return prefix.NewStore(ctx.KVStore(k.storeKey), types.CreateDenomAddressPrefix(denom))
+}
+
+// UnmarshalBalanceCompat unmarshal balance amount from storage, it's backward-compatible with the legacy format.
+func UnmarshalBalanceCompat(cdc codec.BinaryCodec, bz []byte, denom string) (sdk.Coin, error) {
+	amount := math.ZeroInt()
+	if bz == nil {
+		return sdk.NewCoin(denom, amount), nil
+	}
+
+	if err := amount.Unmarshal(bz); err != nil {
+		// try to unmarshal with the legacy format.
+		var balance sdk.Coin
+		if cdc.Unmarshal(bz, &balance) != nil {
+			// return with the original error
+			return sdk.Coin{}, err
+		}
+		return balance, nil
+	}
+
+	return sdk.NewCoin(denom, amount), nil
 }